+++ R/utils.R
@@ -275,6 +275,87 @@ linear_gradient <- function(cols) {
 
 
 
+# shiny utils -------------------------------------------------------------
+
+normalizeChoicesArgs <- function(choices, choiceNames, choiceValues, mustExist = TRUE) {
+  if (is.null(choices)) {
+    if (is.null(choiceNames) || is.null(choiceValues)) {
+      if (mustExist) {
+        stop("Please specify a non-empty vector for `choices` (or, ",
+             "alternatively, for both `choiceNames` AND `choiceValues`).")
+      }
+      else {
+        if (is.null(choiceNames) && is.null(choiceValues)) {
+          return(list(choiceNames = NULL, choiceValues = NULL))
+        }
+        else {
+          stop("One of `choiceNames` or `choiceValues` was set to ",
+               "NULL, but either both or none should be NULL.")
+        }
+      }
+    }
+    if (length(choiceNames) != length(choiceValues)) {
+      stop("`choiceNames` and `choiceValues` must have the same length.")
+    }
+    if (anyNamed(choiceNames) || anyNamed(choiceValues)) {
+      stop("`choiceNames` and `choiceValues` must not be named.")
+    }
+  }
+  else {
+    if (!is.null(choiceNames) || !is.null(choiceValues)) {
+      warning("Using `choices` argument; ignoring `choiceNames` and `choiceValues`.")
+    }
+    choices <- choicesWithNames(choices)
+    choiceNames <- names(choices)
+    choiceValues <- unname(choices)
+  }
+  return(list(choiceNames = as.list(choiceNames), choiceValues = as.list(as.character(choiceValues))))
+}
+
+
+choicesWithNames <- function(choices) {
+  listify <- function(obj) {
+    makeNamed <- function(x) {
+      if (is.null(names(x)))
+        names(x) <- character(length(x))
+      x
+    }
+    res <- lapply(obj, function(val) {
+      if (is.list(val))
+        listify(val)
+      else if (length(val) == 1 && is.null(names(val)))
+        val
+      else makeNamed(as.list(val))
+    })
+    makeNamed(res)
+  }
+  choices <- listify(choices)
+  if (length(choices) == 0)
+    return(choices)
+  choices <- mapply(choices, names(choices), FUN = function(choice,
+                                                            name) {
+    if (!is.list(choice))
+      return(choice)
+    if (name == "")
+      stop("All sub-lists in \"choices\" must be named.")
+    choicesWithNames(choice)
+  }, SIMPLIFY = FALSE)
+  missing <- names(choices) == ""
+  names(choices)[missing] <- as.character(choices)[missing]
+  choices
+}
+
+
+anyNamed <- function(x) {
+  if (length(x) == 0) 
+    return(FALSE)
+  nms <- names(x)
+  if (is.null(nms)) 
+    return(FALSE)
+  any(nzchar(nms))
+}
+
+